Create a flag on Sponsors for Job Fair Participation
Issue #638 and #643 both relate to determining if a Sponsor is participating in the Job Fair or not!

This adds a Boolean that can annotate Job Fair Participation for each sponsor individually and exposes that in the the Django Admin.

PyCon staff can now explicitly set or unset Job Fair Participation so that we can auto generate the list of Job Fair Participants for #643 and hopefully expose the boolean for templates to contain different content for Sponsors who are and are not participating in the job fair.